# Revise message dispatch and primitives