--- a/src/de.hpi.swa.trufflesqueak/src/de/hpi/swa/trufflesqueak/model/AbstractSqueakObject.java
+++ b/src/de.hpi.swa.trufflesqueak/src/de/hpi/swa/trufflesqueak/model/AbstractSqueakObject.java
@@ -145,7 +145,7 @@ public abstract class AbstractSqueakObject implements TruffleObject {
                             try {
                                 return primitiveNode.executeWithArguments(frame, receiver, arguments);
                             } catch (final PrimitiveFailed pf) {
-                                DispatchUtils.handlePrimitiveFailedIndirect(node, primitiveNode, method, pf);
+                                DispatchUtils.handlePrimitiveFailedIndirect(node, method, pf);
                             }
                         }
                     }