… .github/workflows/ci.yml
… mx.trufflesqueak/eclipse-settings/org.eclipse.jdt.core.prefs
… src/de.hpi.swa.trufflesqueak.test/src/de/hpi/swa/trufflesqueak/test/AbstractSqueakTestCase.java
… src/de.hpi.swa.trufflesqueak.test/src/de/hpi/swa/trufflesqueak/test/AbstractSqueakTestCaseWithDummyImage.java
… src/de.hpi.swa.trufflesqueak.test/src/de/hpi/swa/trufflesqueak/test/AbstractSqueakTestCaseWithImage.java
… src/de.hpi.swa.trufflesqueak.test/src/de/hpi/swa/trufflesqueak/test/SqueakBytecodeTest.java
… src/de.hpi.swa.trufflesqueak.test/src/de/hpi/swa/trufflesqueak/test/SqueakPrimitiveTest.java
… src/de.hpi.swa.trufflesqueak/src/de/hpi/swa/trufflesqueak/model/AbstractSqueakObjectWithClassAndHash.java
… src/de.hpi.swa.trufflesqueak/src/de/hpi/swa/trufflesqueak/model/ClassObject.java
… src/de.hpi.swa.trufflesqueak/src/de/hpi/swa/trufflesqueak/model/CompiledCodeObject.java
… src/de.hpi.swa.trufflesqueak/src/de/hpi/swa/trufflesqueak/model/ContextObject.java
… src/de.hpi.swa.trufflesqueak/src/de/hpi/swa/trufflesqueak/model/NativeObject.java
… src/de.hpi.swa.trufflesqueak/src/de/hpi/swa/trufflesqueak/model/layout/ObjectLayouts.java
… src/de.hpi.swa.trufflesqueak/src/de/hpi/swa/trufflesqueak/nodes/AboutToReturnNode.java
… src/de.hpi.swa.trufflesqueak/src/de/hpi/swa/trufflesqueak/nodes/AbstractNode.java
… src/de.hpi.swa.trufflesqueak/src/de/hpi/swa/trufflesqueak/nodes/CacheLimits.java
… src/de.hpi.swa.trufflesqueak/src/de/hpi/swa/trufflesqueak/nodes/DoItRootNode.java
… src/de.hpi.swa.trufflesqueak/src/de/hpi/swa/trufflesqueak/nodes/ExecuteBytecodeNode.java
… src/de.hpi.swa.trufflesqueak/src/de/hpi/swa/trufflesqueak/nodes/ExecuteNonFailingPrimitiveRootNode.java
… src/de.hpi.swa.trufflesqueak/src/de/hpi/swa/trufflesqueak/nodes/ExecuteTopLevelContextNode.java
… src/de.hpi.swa.trufflesqueak/src/de/hpi/swa/trufflesqueak/nodes/HandleNonLocalReturnNode.java
… src/de.hpi.swa.trufflesqueak/src/de/hpi/swa/trufflesqueak/nodes/HandlePrimitiveFailedNode.java
… src/de.hpi.swa.trufflesqueak/src/de/hpi/swa/trufflesqueak/nodes/InheritsFromNode.java
… src/de.hpi.swa.trufflesqueak/src/de/hpi/swa/trufflesqueak/nodes/LookupMethodNode.java
… src/de.hpi.swa.trufflesqueak/src/de/hpi/swa/trufflesqueak/nodes/MaterializeContextOnMethodExitNode.java
… src/de.hpi.swa.trufflesqueak/src/de/hpi/swa/trufflesqueak/nodes/SendSelectorNode.java
… src/de.hpi.swa.trufflesqueak/src/de/hpi/swa/trufflesqueak/nodes/SqueakGuards.java
… src/de.hpi.swa.trufflesqueak/src/de/hpi/swa/trufflesqueak/nodes/accessing/ArrayObjectNodes.java
… src/de.hpi.swa.trufflesqueak/src/de/hpi/swa/trufflesqueak/nodes/bytecodes/AbstractBytecodeNode.java
… src/de.hpi.swa.trufflesqueak/src/de/hpi/swa/trufflesqueak/nodes/bytecodes/AbstractSqueakBytecodeDecoder.java